// src/models/Venta.ts
export interface Venta {
  nombreEstilistaAux?: string;
  User: number;
  id?: number;
  Cia: number;
  Sucursal: number;
  Fecha: string;
  Caja: number;
  No_venta: number;
  Clave_prod: number;
  Cant_producto: number;
  Precio: number;
  Precio_base: number;
  Cve_cliente: number;
  Tasa_iva: number;
  ieps: number;
  Observacion: string;
  Descuento: number;
  Clave_Descuento: number;
  Usuario: number;
  Credito: boolean;
  Corte: number;
  Corte_parcial: number;
  Costo: number;
  cancelada: boolean;
  idEstilista: number;
  folio_estilista: number /* FIJA*/;
  hora: any;
  tiempo: number;
  terminado: boolean;
  validadoServicio: boolean;
  producto?: string;
  estilista?: string;
  cliente?: string;
  d_estilista?: string;
  d_producto?: string;
  d_existencia?: number;
  d_descuento?: number;
  idestilistaAux?: number;
  idRecepcionista?: number;
  d_estilistaAuxilliar?: string;
}

// Generated by https://quicktype.io

// Generated by https://quicktype.io

export interface VentaResponse {
  id: number;
  Cia: number;
  Sucursal: number;
  d_sucursal: string;
  Fecha: string;
  Caja: number;
  No_venta: number;
  no_venta2: number;
  Clave_prod: number;
  d_producto: string;
  Cant_producto: number;
  Precio: number;
  Cve_cliente: number;
  d_cliente: string;
  Tasa_iva: number;
  Observacion: string;
  Descuento: number;
  Clave_Descuento: number;
  User: number;
  d_estilista: null;
  Corte: number;
  Corte_parcial: number;
  Costo: number;
  Precio_base: number;
  No_venta_original: number;
  cancelada: boolean;
  folio_estilista: string;
  hora: string;
  tiempo: number;
  terminado: boolean;
  validadoServicio: boolean;
  idestilistaAux: number;
  idRecepcionista: number;
}